+++ packages/orq-rc/src/models/components/evaluatorresponsellm.ts
@@ -267,8 +267,8 @@ export const EvaluatorResponseLlm$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-07-09T07:44:19.589Z"),
-  updated: z.string().default("2026-07-09T07:44:19.589Z"),
+  created: z.string().default("2026-07-09T12:21:08.426Z"),
+  updated: z.string().default("2026-07-09T12:21:08.426Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.any().optional(),
   type: z.literal("llm_eval"),