+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-09T07:44:18.154Z",
+    "2026-07-09T12:21:06.371Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {