+++ packages/orq-rc/src/models/operations/modelcreate.ts
@@ -55,6 +55,7 @@ export type ModelCreateResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
+  sharing?: components.Config | undefined;
   updated: Date;
 };
 
@@ -151,6 +152,7 @@ export const ModelCreateResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
+  sharing: components.Config$inboundSchema.optional(),
   updated: z.string().datetime({ offset: true }).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {