+++ packages/orq-rc/src/models/operations/modelcreateawsbedrock.ts
@@ -67,6 +67,7 @@ export type ModelCreateAwsBedrockResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
+  sharing?: components.Config | undefined;
   updated: Date;
 };
 
@@ -201,6 +202,7 @@ export const ModelCreateAwsBedrockResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
+  sharing: components.Config$inboundSchema.optional(),
   updated: z.string().datetime({ offset: true }).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {