+++ packages/orq-rc/src/models/operations/modelcreateopenailike.ts
@@ -59,6 +59,7 @@ export type ModelCreateOpenAILikeResponseBody = {
   pricingUrl: string | null;
   provider: string;
   refId: string;
+  sharing?: components.Config | undefined;
   updated: Date;
 };
 
@@ -169,6 +170,7 @@ export const ModelCreateOpenAILikeResponseBody$inboundSchema: z.ZodType<
   pricing_url: z.nullable(z.string()),
   provider: z.string(),
   refId: z.string(),
+  sharing: components.Config$inboundSchema.optional(),
   updated: z.string().datetime({ offset: true }).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {